foreign_stock/requirements.txt
aiohttp==3.9.3
asyncpg==0.29.0
discord.py==2.3.2
python-dotenv==1.0.1
